Added melody- and harmony- specific subclasses to parse differently for the different neural nets.
This is needed since the representation needed for the two networks differs in some respects: For the melody neural net, we need the accompanying chords as the midi pitch heights for each note. For the harmony neural net, on the other hand, we only need the chord once and independent of the notes played beneath, i.e. the representation is much more sparse. Therefore, two different subclasses of the SongParser were written.