Enhanced and rewrote musical element classes.
Rewrote the different Midi-classes into one coherent class for chords and notes, respectively.
The Note and Chord classes both have different class methods that act as constructor wrappers now.
For example, one can instantiate a chord as a pitch height, a symbol, or a neural net representation,
respectively - depending on the constructor wrapper used. This facilitates the work with the classes.