Added small improvements to the song class.
Changede naming of one variable; made sure that all offsets are integers.
Additionally, ensured that the offsets also work for chords that are played over
one note (here, the offset is with respect to the previous note; therefore, we need
to add this offset to the overall offset to get the final offset).