Added tests for the MelodyParser class and refactored one test into tests for parent class.
The tests mostly check that the parsing and the conversion to the neural net input work as expected.
The refactored test is the one that checks the "sliding_window" method which is general method of the
class that both child classes inherit; therefore, this tests belongs to the tests for the parent class.