Rewrote the pipeline class to work with a dictionary as the config file.
This makes experimentation much easier: One can simply pass a .yml file with all the relevant information
and the training process starts. In the .yml file, everything about the whole training process (the input
size, the model architecture, the losses etc.) can be specified; therefore, the whole training process
can be controlled with this one file.